# small optimizations.