--- a/decode.go
+++ b/decode.go
@@ -18,12 +18,6 @@ const (
 	disallowUnknownFieldsFlag
 )
 
-const (
-	bytesAllocLimit = 1e6 // 1mb
-	sliceAllocLimit = 1e4
-	maxMapSize      = 1e6
-)
-
 type bufReader interface {
 	io.Reader
 	io.ByteScanner
@@ -622,37 +616,19 @@ func readN(r io.Reader, b []byte, n int) ([]byte, error) {
 		switch {
 		case n < 64:
 			b = make([]byte, 0, 64)
-		case n <= bytesAllocLimit:
-			b = make([]byte, 0, n)
 		default:
-			b = make([]byte, 0, bytesAllocLimit)
+			b = make([]byte, 0, n)
 		}
 	}
 
-	if n <= cap(b) {
+	if n > cap(b) {
+		b = append(b, make([]byte, n-len(b))...)
+	} else if n <= cap(b) {
 		b = b[:n]
-		_, err := io.ReadFull(r, b)
-		return b, err
-	}
-	b = b[:cap(b)]
-
-	var pos int
-	for {
-		alloc := min(n-len(b), bytesAllocLimit)
-		b = append(b, make([]byte, alloc)...)
-
-		_, err := io.ReadFull(r, b[pos:])
-		if err != nil {
-			return b, err
-		}
-
-		if len(b) == n {
-			break
-		}
-		pos = len(b)
 	}
 
-	return b, nil
+	_, err := io.ReadFull(r, b)
+	return b, err
 }
 
 func min(a, b int) int { //nolint:unparam