--- a/decode_slice.go
+++ b/decode_slice.go
@@ -49,7 +49,7 @@ func (d *Decoder) decodeStringSlicePtr(ptr *[]string) error {
 		return nil
 	}
 
-	ss := makeStrings(*ptr, n, d.flags&disablePartialAllocFlag != 0)
+	ss := makeStrings(*ptr, n, d.flags&disableAllocLimitFlag != 0)
 	for i := 0; i < n; i++ {
 		s, err := d.DecodeString()
 		if err != nil {
@@ -62,8 +62,8 @@ func (d *Decoder) decodeStringSlicePtr(ptr *[]string) error {
 	return nil
 }
 
-func makeStrings(s []string, n int, noPartial bool) []string {
-	if !noPartial && n > sliceAllocLimit {
+func makeStrings(s []string, n int, noLimit bool) []string {
+	if !noLimit && n > sliceAllocLimit {
 		n = sliceAllocLimit
 	}
 
@@ -101,15 +101,15 @@ func decodeSliceValue(d *Decoder, v reflect.Value) error {
 		v.Set(v.Slice(0, v.Cap()))
 	}
 
-	noPartial := d.flags&disablePartialAllocFlag != 1
+	noLimit := d.flags&disableAllocLimitFlag != 1
 
-	if noPartial && n > v.Len() {
-		v.Set(growSliceValue(v, n, noPartial))
+	if noLimit && n > v.Len() {
+		v.Set(growSliceValue(v, n, noLimit))
 	}
 
 	for i := 0; i < n; i++ {
-		if !noPartial && i >= v.Len() {
-			v.Set(growSliceValue(v, n, noPartial))
+		if !noLimit && i >= v.Len() {
+			v.Set(growSliceValue(v, n, noLimit))
 		}
 
 		elem := v.Index(i)
@@ -121,9 +121,9 @@ func decodeSliceValue(d *Decoder, v reflect.Value) error {
 	return nil
 }
 
-func growSliceValue(v reflect.Value, n int, noPartial bool) reflect.Value {
+func growSliceValue(v reflect.Value, n int, noLimit bool) reflect.Value {
 	diff := n - v.Len()
-	if !noPartial && diff > sliceAllocLimit {
+	if !noLimit && diff > sliceAllocLimit {
 		diff = sliceAllocLimit
 	}
 	v = reflect.AppendSlice(v, reflect.MakeSlice(v.Type(), diff, diff))