--- a/encode.go
+++ b/encode.go
@@ -124,6 +124,8 @@ func (e *Encoder) WithDict(dict map[string]int, fn func(*Encoder) error) error {
 func (e *Encoder) ResetWriter(w io.Writer) {
 	if bw, ok := w.(writer); ok {
 		e.w = bw
+	} else if w == nil {
+		e.w = nil
 	} else {
 		e.w = newByteWriter(w)
 	}