--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,14 @@
 module github.com/vmihailenco/msgpack/v5
 
-go 1.11
+go 1.19
 
 require (
 	github.com/stretchr/testify v1.6.1
 	github.com/vmihailenco/tagparser/v2 v2.0.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )